pyproject.toml
[project]
name = "phantom-seed"
version = "0.1.0"
description = "A roguelike visual novel with real-time AI-generated content"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pygame-ce>=2.5",
    "openai>=1.0",
    "httpx>=0.28",
    "pydantic>=2.10",
    "pillow>=11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[project.scripts]
phantom-seed = "phantom_seed.main:main"

[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"
